requirements.txt
streamlit==1.51.0
pandas==2.2.3
matplotlib==3.10.1 
pdfplumber==0.11.6
scikit-learn==1.6.1
pydantic==2.9.2
psycopg2-binary==2.9.10
python-dotenv==1.1.1
altair==5.5.0
passlib[argon2]==1.7.4